Permisos Usuarios, Socios y Disciplinas completos

+++ admin/src/web/controllers/disciplinas.py
@@ -1,6 +1,7 @@
 import json
-from flask import Blueprint, render_template, request, redirect, flash, session
+from flask import Blueprint, render_template, request, redirect, flash, session, abort
 from src.core import disciplinas
+from src.web.helpers.permission import check_permission
 from src.decoradores.login import login_requerido
 from src.core import usuarios
 
@@ -17,20 +18,26 @@ def disciplina_json():
 @login_requerido
 def disciplina_index():
     """Muestra las disciplinas de la página indicada en el request. Si no hay request, la página será la primera"""
-    page = request.args.get("page", 1, type=int)
-    kwargs = {
-        "disciplinas": disciplinas.listar_disciplinas(page),
-        "usuario": usuarios.buscar_usuario_email(session["user"]),
-    }
-    return render_template("disciplinas/index.html", **kwargs)
+    if (check_permission(session["user"], "disciplina_index")):
+        page = request.args.get("page", 1, type=int)
+        kwargs = {
+            "disciplinas": disciplinas.listar_disciplinas(page),
+            "usuario": usuarios.buscar_usuario_email(session["user"]),
+        }
+        return render_template("disciplinas/index.html", **kwargs)
+    else:
+        return abort(403)
 
 
 @disciplina_blueprint.route("/alta-disciplina")
 @login_requerido
 def form_disciplina():
     """Devuelve el template con el formulario para agregar una disciplina"""
-    kwargs = {"usuario": usuarios.buscar_usuario_email(session["user"])}
-    return render_template("disciplinas/alta_disciplinas.html", **kwargs)
+    if (check_permission(session["user"], "disciplina_new")):
+        kwargs = {"usuario": usuarios.buscar_usuario_email(session["user"])}
+        return render_template("disciplinas/alta_disciplinas.html", **kwargs)
+    else:
+        return abort(403)
 
 
 @disciplina_blueprint.route("/<id>")
@@ -76,37 +83,43 @@ def disciplina_add():
 @login_requerido
 def disciplina_update():
     """Llama a las funciones del modelo para validar los inputs del formulario para modificar una disciplina. Si los inputs son validos, le dice al modelo que la modifique"""
-    data_disciplina = {
-        "id": request.form.get("id"),
-        "nombre": request.form.get("nombre").capitalize(),
-        "categoria": request.form.get("categoria").capitalize(),
-        "instructores": request.form.get("instructores"),
-        "horarios": request.form.get("horarios"),
-        "costo": request.form.get("costo"),
-        "habilitada": (request.form.get("habilitada") == "Si"),
-    }
-    resultado, mensaje = disciplinas.validar_inputs(data_disciplina)
-    if resultado:
-        resultado, mensaje = disciplinas.validar_disciplina_repetida(
-            data_disciplina["nombre"],
-            data_disciplina["categoria"],
-            "modificacion",
-            data_disciplina["id"],
-        )
+    if (check_permission(session["user"], "disciplina_update")):
+        data_disciplina = {
+            "id": request.form.get("id"),
+            "nombre": request.form.get("nombre").capitalize(),
+            "categoria": request.form.get("categoria").capitalize(),
+            "instructores": request.form.get("instructores"),
+            "horarios": request.form.get("horarios"),
+            "costo": request.form.get("costo"),
+            "habilitada": (request.form.get("habilitada") == "Si"),
+        }
+        resultado, mensaje = disciplinas.validar_inputs(data_disciplina)
         if resultado:
-            disciplinas.modificar_disciplina(data_disciplina)
-            return redirect("/disciplinas")
+            resultado, mensaje = disciplinas.validar_disciplina_repetida(
+                data_disciplina["nombre"],
+                data_disciplina["categoria"],
+                "modificacion",
+                data_disciplina["id"],
+            )
+            if resultado:
+                disciplinas.modificar_disciplina(data_disciplina)
+                return redirect("/disciplinas")
+            else:
+                flash(mensaje)
+                return redirect("/disciplinas/" + data_disciplina["id"])
         else:
             flash(mensaje)
             return redirect("/disciplinas/" + data_disciplina["id"])
     else:
-        flash(mensaje)
-        return redirect("/disciplinas/" + data_disciplina["id"])
+        return abort(403)
 
 
 @disciplina_blueprint.route("/eliminar/<id>", methods=["DELETE", "GET"])
 @login_requerido
 def disciplina_delete(id):
     """Le dice al modelo que borre la disciplina enviada"""
-    disciplinas.eliminar_disciplina(id)
-    return redirect("/disciplinas")
+    if (check_permission(session["user"], "disciplina_destroy")):
+        disciplinas.eliminar_disciplina(id)
+        return redirect("/disciplinas")
+    else:
+        return abort(403)
